requirements.txt
bintrees==2.0.7
requests>=2.20.0
six==1.10.0
websocket-client==0.40.0
pymongo==3.5.1
pytest>=3.3.0
pytest-cov>=2.5.0
jsonpickle
plotly
python-binance
gdax
krakenex
python-kucoin
python-dateutil>=2.8.0
